Optimize trim function
If [range] is excluded, ":{range}call" will call the function for each line in
the range, with the cursor on the start of each line.

The original implementation has performance issue which process the entire range
for each line.

We fix and optimize it by eliminating the for loop and remain the lines
unchanged if nothing has been trimmed.